Add more comments, move to `ioutil` package, switch to using teeReader for copying bytes to the buffer.
