+++ internal/ioutil/rereader.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package rereader
+package ioutil
 
 import (
 	"bytes"
@@ -37,17 +37,17 @@ import (
 //   reset we should flush the rest of the reader into the buffer, or
 //   combine the buffer and source reader together dynamically.
 
-// New will return a new ReReader and a closer to cleanup the ReReader once
-// the callsite is done with it.
-func New(source io.Reader) (*ReReader, func()) {
+// NewRereader returns a new re-reader and closer that wraps the given reader.
+// The re-reader consumes the given reader on demand, recording the entire
+// stream so it can be replayed. The re-reader is suitable for retries. The
+// re-reader does not support concurrent consumers, as would be necessary
+// for speculative retries.
+func NewRereader(src io.Reader) (*ReReader, func()) {
 	buf := buffer.Get()
-	closer := func() {
-		buffer.Put(buf)
-	}
 	return &ReReader{
-		src: source,
-		buf: buf,
-	}, closer
+		teeReader: io.TeeReader(src, buf),
+		buf:       buf,
+	}, func() { buffer.Put(buf) }
 }
 
 // ReReader has the ability to read the same io.Reader multiple times.
@@ -55,45 +55,56 @@ func New(source io.Reader) (*ReReader, func()) {
 //   - Reset MUST be called in order to Re-Read the reader.
 //   - Reset MUST be called after the ReReader is exausted (Read will return
 //     an io.EOF error).
+//   - Concurrent reads are not supported.
 type ReReader struct {
-	src    io.Reader
-	buf    *bytes.Buffer
-	reader *bytes.Reader
+	// teeReader is an io.TeeReader that will read from the source io.Reader
+	// and simultaneously return the result and write all the data to the
+	// buf attribute to be replayed.
+	teeReader io.Reader
 
-	// Small buffer required to test whether the reader is exausted (EOF).
+	// buf will be filled with the contents from the teeReader as we read
+	// from the tee reader.
+	buf *bytes.Buffer
+
+	// Unfortunately we can't reset a bytes.Buffer to reread from it, so after
+	// our initial read into the buffer, we need to use a reader to read the
+	// bytes out of the buffer.
+	bufReader *bytes.Reader
+
+	// In order to test whether the reader is finished, we need to call `Read`
+	// and validate that it returns an io.EOF error.  This buffer is used so
+	// we don't need to reallocate a slice for every `Read` check.
 	testbuf [1]byte
 
+	// useBuffer indicates whether reads should go to the teeReader or the
+	// bufReader.
 	useBuffer atomic.Bool
 }
 
-// Read implements the io.Reader interface.
+// Read implements the io.Reader interface.  On the first read, we will record
+// the entire contents of the source io.Reader for subsequent Reads.
 func (rr *ReReader) Read(p []byte) (n int, err error) {
 	if rr.useBuffer.Load() {
-		return rr.reader.Read(p)
+		return rr.bufReader.Read(p)
 	}
 
-	n, err = rr.src.Read(p)
-	if n > 0 {
-		if n, err := rr.buf.Write(p[:n]); err != nil {
-			return n, err
-		}
-	}
-	return
+	return rr.teeReader.Read(p)
 }
 
-// Reset validates that the current Reader is finished (EOF) before resetting
-// the Reader to read from the beginning of the source reader again.
+// Reset resets the rereader to read from the beginning of the source data.
+//   - If the current Read is not finished (a call to `Read` does not
+//     return an io.EOF error), Reset will return an error.
 func (rr *ReReader) Reset() error {
 	if _, err := rr.Read(rr.testbuf[:]); err != io.EOF {
 		return errors.New("cannot reset the rereader until we've finished reading the current reader")
 	}
 
 	if !rr.useBuffer.Load() {
-		rr.reader = bytes.NewReader(rr.buf.Bytes())
+		rr.bufReader = bytes.NewReader(rr.buf.Bytes())
 		rr.useBuffer.Store(true)
 		return nil
 	}
 
-	_, err := rr.reader.Seek(0, io.SeekStart)
+	_, err := rr.bufReader.Seek(0, io.SeekStart)
 	return err
 }
